Move ValueExporter to ProfileData
